Added functions to make moves on the board
You can plug moves directly from get_legal_moves into these functions
and it should work. Some work is needed requiring the setting of the
next_play variable, as it should be set to 0 if move[2] is won. This is
one of the places where preventing moves on won boards should be done.

Signed-off-by: ZweiSpeedruns <zcm727@gmail.com>